Move JVersion.h out of CLI
