+++ CMakeLists.txt
@@ -194,6 +194,7 @@ message(STATUS "-----------------------")
 
 include_directories(src/libraries)   # So that everyone can find the JANA header files
 include_directories(src/external)    # So that everyone can find our vendorized header-old libraries
+include_directories(${CMAKE_CURRENT_BINARY_DIR}/src/libraries)   # So that everyone can find JVersion.h
 
 # This is needed on macos to allow plugins to link without resolving all JANA symbols until runtime
 if (${CMAKE_SYSTEM_NAME} MATCHES "Darwin")
