+++ src/libraries/JANA/CLI/JVersion.h
@@ -1,0 +1,8 @@
+#pragma once
+
+// We are moving JVersion.h out of CLI. 
+// In the future, use #include <JANA/JVersion.h> instead.
+
+#include <JANA/JVersion.h>
+
+
