+++ src/libraries/JANA/Compatibility/JGeometryXML.h
@@ -12,7 +12,7 @@
 #include <JANA/Compatibility/JGeometry.h>
 #include <JANA/Compatibility/JStreamLog.h>
 #include <JANA/Calibrations/JCalibration.h>
-#include <JANA/CLI/JVersion.h>
+#include <JANA/JVersion.h>
 
 #include <JANA/Compatibility/md5.h>
