+++ src/python/modules/jana/jana_module.cc
@@ -4,7 +4,7 @@
 
 #include <janapy.h>
 #include <JANA/CLI/JMain.h>
-#include <JANA/CLI/JVersion.h>
+#include <JANA/JVersion.h>
 
 // Something to throw that makes a nicer error message
 class PYTHON_MODULE_STARTUP_FAILED{public: PYTHON_MODULE_STARTUP_FAILED(){}};
